+++ packages/api/src/defaults/variants/index.ts
@@ -1,3 +1,4 @@
 export * from "./mediaQuery";
 export * from "./opacity";
 export * from "./pseudoClass";
+export * from "./pseudoElement";