+++ packages/api/src/defaults/variants/preset.ts
@@ -1,7 +1,13 @@
-import { mediaQueries, opacity, pseudoClasses } from "@/defaults/variants";
+import {
+	mediaQueries,
+	opacity,
+	pseudoClasses,
+	pseudoElements,
+} from "@/defaults/variants";
 
 export const standardPreset = {
 	pseudoClasses,
+	pseudoElements,
 	mediaQueries,
 };
 
@@ -9,4 +15,5 @@ export const opacityPreset = {
 	mediaQueries,
 	opacity,
 	pseudoClasses,
+	pseudoElements,
 };